# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления арендой автомобилей (CRM).",
        "title": "Auto Hub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Без фильтра возвращает UNPAID + OVERDUE",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Список платежей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PAID, UNPAID, OVERDUE, PENDING (=UNPAID+OVERDUE)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaymentsListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отмечает платёж как PAID и пересчитывает paid_amount договора\nНельзя удалить оплаченный платёж",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "payments",
                    "payments"
                ],
                "summary": "Удалить платёж",
                "parameters": [
                    {
                        "description": "ID платежа и дата оплаты",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PayPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PayPaymentResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Платёж уже оплачен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отмечает платёж как PAID и пересчитывает paid_amount договора\nНельзя удалить оплаченный платёж",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "payments",
                    "payments"
                ],
                "summary": "Удалить платёж",
                "parameters": [
                    {
                        "description": "ID платежа и дата оплаты",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PayPaymentRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID платежа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PayPaymentResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Платёж уже оплачен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                }
            }
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "PAID",
                "UNPAID",
                "OVERDUE"
            ],
            "x-enum-varnames": [
                "PaymentStatusPaid",
                "PaymentStatusUnpaid",
                "PaymentStatusOverdue"
            ]
        },
...
        "dto.CarForPayment": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                }
            }
        },
...
        "dto.ContractForPayment": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/dto.CarForPayment"
                },
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/dto.DriverForPayment"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ContractSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "paid_amount": {
                    "type": "number"
                },
                "remaining_balance": {
                    "type": "number"
                }
            }
        },
...
        "dto.DriverForPayment": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.PayPaymentRequest": {
            "type": "object",
            "required": [
                "payment_id"
            ],
            "properties": {
                "paid_at": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                }
            }
        },
        "dto.PayPaymentResponse": {
            "type": "object",
            "properties": {
                "contract": {
                    "$ref": "#/definitions/dto.ContractSummary"
                },
                "payment": {
                    "$ref": "#/definitions/domain.Payment"
                }
            }
        },
...
        "dto.PaymentWithContract": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contract": {
                    "$ref": "#/definitions/dto.ContractForPayment"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                }
            }
        },
        "dto.PaymentsListResponse": {
            "type": "object",
            "properties": {
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PaymentWithContract"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}